+++ components/ui/animated-popper/index.ts
@@ -1,0 +1,17 @@
+'use client';
+import dynamic from 'next/dynamic';
+
+const AnimatedPopperContent = dynamic(
+  () => import('./animated-popper-content'),
+  {
+    ssr: false,
+  },
+);
+
+export {
+  AnimatedPopperOverlay,
+  AnimatedPopperTrigger,
+  AnimatedPopperProvider,
+  AnimatedPopper,
+} from './animated-popper';
+export { AnimatedPopperContent };